Create Nmap-Discovery-Scan.py
This will do a TCP SYN "ping" across IP addresses and/or Subnets kept in a list in a .txt file. It then outputs what hosts are in the up state into a CSV file. 
You need to make sure you install the python-nmap module to be imported by the script.